+++ R/estimate_grouplevel.R
@@ -43,14 +43,11 @@
 #' # Visualize random effects
 #' plot(random)
 #'
-#' # Reshape to wide data so that it matches the original dataframe...
-#' reshaped <- reshape_grouplevel(random, indices = c("Coefficient", "SE"))
+#' # Reshape to wide data...
+#' reshaped <- reshape_grouplevel(random, group = "carb", indices = c("Coefficient", "SE"))
 #'
 #' # ...and can be easily combined with the original data
-#' alldata <- cbind(mtcars, reshaped)
-#'
-#' # Use summary() to remove duplicated rows
-#' summary(reshaped)
+#' alldata <- merge(mtcars, reshaped)
 #'
 #' # overall coefficients
 #' estimate_grouplevel(model, type = "total")
@@ -76,8 +73,7 @@ estimate_grouplevel.default <- function(model,
   # Extract params
   params <- parameters::model_parameters(
     model,
-    effects = ifelse(type == "random", "all", "total"),
-    group_level = identical(type, "random"),
+    effects = ifelse(type == "random", "grouplevel", "total"),
     ...
   )
 
@@ -93,7 +89,7 @@ estimate_grouplevel.default <- function(model,
   }
 
   # TODO: improve / add new printing that groups by group/level?
-  random <- as.data.frame(params[params$Effects == type, ])
+  random <- as.data.frame(params)
 
   # Remove columns with only NaNs (as these are probably those of fixed effects)
   random[vapply(random, function(x) all(is.na(x)), TRUE)] <- NULL
@@ -128,8 +124,7 @@ estimate_grouplevel.brmsfit <- function(model,
   # Extract params
   params <- parameters::model_parameters(
     model,
-    effects = ifelse(type == "random", "all", "total"),
-    group_level = identical(type, "random"),
+    effects = ifelse(type == "random", "grouplevel", "total"),
     dispersion = dispersion,
     test = test,
     diagnostic = diagnostic,
@@ -138,6 +133,11 @@ estimate_grouplevel.brmsfit <- function(model,
 
   # get cleaned parameter names with additional information
   clean_parameters <- attributes(params)$clean_parameters
+
+  # match parameters
+  if (!is.null(clean_parameters)) {
+    clean_parameters <- clean_parameters[clean_parameters$Parameter %in% params$Parameter, , drop = FALSE]
+  }
 
   # Re-add info
   if (!"Group" %in% names(params) && !is.null(clean_parameters)) {
@@ -148,7 +148,7 @@ estimate_grouplevel.brmsfit <- function(model,
   }
 
   # TODO: improve / add new printing that groups by group/level?
-  random <- as.data.frame(params[params$Effects == type, ])
+  random <- as.data.frame(params)
 
   # Remove columns with only NaNs (as these are probably those of fixed effects)
   random[vapply(random, function(x) all(is.na(x)), TRUE)] <- NULL
@@ -206,8 +206,7 @@ estimate_grouplevel.stanreg <- function(model,
   # Extract params
   params <- parameters::model_parameters(
     model,
-    effects = ifelse(type == "random", "all", "total"),
-    group_level = identical(type, "random"),
+    effects = ifelse(type == "random", "grouplevel", "total"),
     dispersion = dispersion,
     test = test,
     diagnostic = diagnostic,
@@ -224,7 +223,12 @@ estimate_grouplevel.stanreg <- function(model,
   if (!is.null(clean_parameters)) {
     # fix for rstanarm, which contains a sigma columns
     clean_parameters <- clean_parameters[
-      clean_parameters$Component != "sigma" & !startsWith(clean_parameters$Parameter, "Sigma["), # nolint
+      clean_parameters$Component != "sigma" & !startsWith(clean_parameters$Parameter, "Sigma["), ,
+      drop = FALSE # nolint
+    ]
+    clean_parameters <- clean_parameters[
+      clean_parameters$Parameter %in% params$Parameter, ,
+      drop = FALSE
     ]
 
     params$Parameter <- insight::trim_ws(sub(":.*", "", clean_parameters$Group))
@@ -233,7 +237,7 @@ estimate_grouplevel.stanreg <- function(model,
   }
 
   # TODO: improve / add new printing that groups by group/level?
-  random <- as.data.frame(params[params$Effects == type, ])
+  random <- as.data.frame(params)
 
   # Remove columns with only NaNs (as these are probably those of fixed effects)
   random[vapply(random, function(x) all(is.na(x)), TRUE)] <- NULL
@@ -270,11 +274,9 @@ estimate_grouplevel.stanreg <- function(model,
 .clean_grouplevel <- function(random) {
   row.names(random) <- NULL
   random$Effects <- NULL
-  if (
-    "Component" %in%
-      names(random) &&
-      insight::has_single_value(random$Component, remove_na = TRUE) &&
-      unique(random$Component) == "conditional"
+  if ("Component" %in% names(random) &&
+    insight::has_single_value(random$Component, remove_na = TRUE) &&
+    unique(random$Component) == "conditional"
   ) {
     random$Component <- NULL
   }