+++ tests/testthat/test-keep_iterations.R
@@ -15,8 +15,8 @@ test_that("estimate_means() - posterior draws", {
     c(
       "names", "class", "row.names", "at", "by", "focal_terms", "adjusted_for",
       "predict", "estimate", "transform", "datagrid", "preserve_range",
-      "model_info", "keep_iterations", "table_title", "table_footer", "model",
-      "response", "ci", "backend", "coef_name"
+      "model_info", "keep_iterations", "joint_test", "table_title",
+      "table_footer", "model", "response", "ci", "backend", "coef_name"
     )
   )
   expect_named(
@@ -38,8 +38,8 @@ test_that("estimate_means() - posterior draws", {
     c(
       "names", "class", "row.names", "at", "by", "focal_terms", "adjusted_for",
       "predict", "estimate", "transform", "datagrid", "preserve_range",
-      "model_info", "keep_iterations", "table_title", "table_footer", "model",
-      "response", "ci", "backend", "coef_name"
+      "model_info", "keep_iterations", "joint_test", "table_title",
+      "table_footer", "model", "response", "ci", "backend", "coef_name"
     )
   )
   expect_named(
@@ -64,7 +64,7 @@ test_that("estimate_contrasts() - posterior draws", {
       "model", "response", "ci", "p_adjust", "backend", "focal_terms",
       "adjusted_for", "predict", "comparison", "contrast", "estimate",
       "transform", "datagrid", "preserve_range", "coef_name", "model_info",
-      "keep_iterations"
+      "keep_iterations", "joint_test"
     )
   )
   expect_named(
@@ -88,7 +88,7 @@ test_that("estimate_contrasts() - posterior draws", {
       "model", "response", "ci", "p_adjust", "backend", "focal_terms",
       "adjusted_for", "predict", "comparison", "contrast", "estimate",
       "transform", "datagrid", "preserve_range", "coef_name", "model_info",
-      "keep_iterations"
+      "keep_iterations", "joint_test"
     )
   )
   expect_named(
@@ -187,7 +187,8 @@ test_that("estimate_contrasts() - posterior draws, emmeans", {
     c(
       "names", "class", "row.names", "table_title", "table_footer",
       "model", "response", "ci", "p_adjust", "backend", "at", "by",
-      "predict", "comparison", "contrast", "transform", "keep_iterations"
+      "predict", "comparison", "contrast", "transform", "keep_iterations",
+      "joint_test"
     )
   )
   expect_named(
@@ -205,7 +206,8 @@ test_that("estimate_contrasts() - posterior draws, emmeans", {
     c(
       "names", "class", "row.names", "table_title", "table_footer",
       "model", "response", "ci", "p_adjust", "backend", "at", "by",
-      "predict", "comparison", "contrast", "transform", "keep_iterations"
+      "predict", "comparison", "contrast", "transform", "keep_iterations",
+      "joint_test"
     )
   )
   expect_identical(dim(out), c(3L, 4007L))
@@ -253,7 +255,7 @@ test_that("estimate_slopes() - posterior draws, get_predicted", {
   expect_named(
     attributes(out),
     c(
-      "names", "row.names", "class", "ci", "keep_iterations", "response",
+      "names", "class", "row.names", "ci", "keep_iterations", "response",
       "transform", "model", "datagrid", "focal_terms", "preserve_range",
       "table_title", "coef_name", "model_info", "table_footer", "adjusted_for",
       "at_specs", "at", "by", "reference", "data"