+++ NEWS.md
@@ -1,9 +1,27 @@
-# modelbased (devel)
+# modelbased 0.11.1
+
+## Changes
+
+* The `comparison` argument can now also be a custom function, or a matrix
+  (e.g., to define contrasts).
+
+* The `comparison` argument can now also be `"joint"`, to jointly test
+  hypotheses (i.e. conducting a joint test) in factorial designs.
+
+* New vignette about user-defined contrasts and joint tests in
+  `estimate_contrasts()`.
+
+# modelbased 0.11.0
 
 ## New functions
 
 * Added `pool_slopes()`, to pool results from `estimate_slopes()` applied to
   imputed data.
+
+## Breaking Changes
+
+* `reshape_grouplevel()` now takes the correct number of specified random effects
+  groups into account when reshaping results.
 
 ## Changes
 
@@ -17,6 +35,10 @@
 
 * `estimate_prediction()` and `estimate_relation()` now support Wiener-models
   (Drift Diffusion Models) from package *brms*.
+
+* `estimate_prediction()`, `estimate_relation()` and similar functions now
+  include the `Row` column for models with ordinal or categorical response
+  variables when the `data` argument was provided.
 
 * `estimate_slopes()` can now also calculate average marginal effects of a
   predictor, just for the trend of that predictor within a certain range of
@@ -35,6 +57,20 @@
 
 * Exposed `iterations` argument in `estimate_prediction()` and `estimate_relation()`.
 
+* Option `estimate = "average` no longer prints information on averaged predictors
+  in the footer, because strictly, the *predictions* are averaged over, and not
+  the non-focal variables.
+
+* Better handling for models with offsets in `estimate_means()` and
+  `estimate_contrasts()`. Informative messages are given when models include
+  offset terms, and it is possible to fix the offset value using the `offset`
+  argument. The `offset` argument is also available for `estimate_relation()`,
+  `estimate_prediction()` and similar.
+
+* For consistency, `estimate_slopes()` now also uses the residual degrees of
+  freedom by default (like `estimate_means()`) when calculating confidence
+  intervals and p-values.
+
 * Minor improvements to the documentation.
 
 ## Bug fixes
@@ -50,6 +86,9 @@
 * Fixed issue with wrong y-axis label for plots from `estimate_slopes()`.
 
 * Fixed issue with weights in `estimate_relation()`.
+
+* Fixed issue in printed output for the statistic column, which should be `z`
+  for the `marginaleffects` backend, when argument `df = Inf`.
 
 # modelbased 0.10.0